Update test cases for PHP 8
The output format for `var_dump()` changed, making it much more difficult to
manipulate the output for an array to make it look like a Ds class. Instead, we
can use `print_r`, which is a little better suited to the task because it
doesn't print out instance ids, and it uses a consistent syntax between arrays
and objects. It's possible, however, that it introduces some false negatives
for test cases that rely on being able to see which values are references in
`var_dump()` output.

Additionally, the Xdebug setting `xdebug.overload_var_dump` went away, but we
can still disable color with the `cli_color` option, however, this may be moot
with the use of `print_r()`.